requirements.txt
accelerate==0.15.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
boto3==1.26.59
botocore==1.29.59
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
diffusers @ git+https://github.com/huggingface/diffusers.git@125d783076e5bd9785beb05367a2d2566843a271
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
filelock==3.9.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.12.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==0.4.3
numpy==1.24.1
orjson==3.8.5
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
psutil==5.9.4
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
runpod==0.8.3
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
tqdm==4.64.1
transformers==4.26.0
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
zipp==3.12.0
